# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/summaries/{summary_id}/topics": {
      "post": {
        "tags": [
          "summaries"
        ],
        "summary": "Create Topic Endpoint",
        "description": "\u8981\u7d04\u306b\u8b70\u984c\u30921\u4ef6\u8ffd\u52a0\u3059\u308b. \u8981\u7d04\u304c\u898b\u3064\u304b\u3089\u306a\u3051\u308c\u3070 404.\n\nArgs:\n    summary_id (UUID): \u89aa\u8981\u7d04\u306e ID.\n    body (TopicCreate): \u8ffd\u52a0\u3059\u308b\u8b70\u984c.\n    db_session (AsyncSession): SQLAlchemy \u306e\u975e\u540c\u671f\u30bb\u30c3\u30b7\u30e7\u30f3.\n    user (User): \u73fe\u5728\u306e\u30e6\u30fc\u30b6\u30fc.\n\nReturns:\n    TopicResponse: \u8ffd\u52a0\u3055\u308c\u305f\u8b70\u984c (id \u4ed8\u304d).\n\nRaises:\n    HTTPException: 404 - \u8981\u7d04\u304c\u898b\u3064\u304b\u3089\u306a\u3044\u5834\u5408.",
        "operationId": "create_topic_endpoint_api_v1_summaries__summary_id__topics_post",
        "parameters": [
          {
            "name": "summary_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Summary Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TopicCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TopicResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/summaries/{summary_id}/topics/{topic_id}": {
      "patch": {
        "tags": [
          "summaries"
        ],
        "summary": "Update Topic Endpoint",
        "description": "\u8b70\u984c\u3092\u90e8\u5206\u66f4\u65b0\u3059\u308b. \u898b\u3064\u304b\u3089\u306a\u3051\u308c\u3070 404, \u5fc5\u9808\u3092 null \u306b\u3059\u308b\u3068 422.\n\nArgs:\n    summary_id (UUID): \u89aa\u8981\u7d04\u306e ID.\n    topic_id (int): \u8b70\u984c\u306e ID.\n    body (TopicEdit): \u66f4\u65b0\u5185\u5bb9.\n    db_session (AsyncSession): SQLAlchemy \u306e\u975e\u540c\u671f\u30bb\u30c3\u30b7\u30e7\u30f3.\n    user (User): \u73fe\u5728\u306e\u30e6\u30fc\u30b6\u30fc.\n\nReturns:\n    TopicResponse: \u66f4\u65b0\u5f8c\u306e\u8b70\u984c.\n\nRaises:\n    HTTPException: 404 - \u8b70\u984c\u304c\u898b\u3064\u304b\u3089\u306a\u3044\u5834\u5408. 422 - \u5fc5\u9808\u30d5\u30a3\u30fc\u30eb\u30c9\u3092 null \u306b\u3057\u305f\u5834\u5408.",
        "operationId": "update_topic_endpoint_api_v1_summaries__summary_id__topics__topic_id__patch",
        "parameters": [
          {
            "name": "summary_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Summary Id"
            }
          },
          {
            "name": "topic_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Topic Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TopicEdit"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TopicResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "summaries"
        ],
        "summary": "Delete Topic Endpoint",
        "description": "\u8b70\u984c\u30921\u4ef6\u524a\u9664\u3059\u308b. \u898b\u3064\u304b\u3089\u306a\u3051\u308c\u3070 404.\n\nArgs:\n    summary_id (UUID): \u89aa\u8981\u7d04\u306e ID.\n    topic_id (int): \u8b70\u984c\u306e ID.\n    db_session (AsyncSession): SQLAlchemy \u306e\u975e\u540c\u671f\u30bb\u30c3\u30b7\u30e7\u30f3.\n    user (User): \u73fe\u5728\u306e\u30e6\u30fc\u30b6\u30fc.\n\nRaises:\n    HTTPException: 404 - \u8b70\u984c\u304c\u898b\u3064\u304b\u3089\u306a\u3044\u5834\u5408.",
        "operationId": "delete_topic_endpoint_api_v1_summaries__summary_id__topics__topic_id__delete",
        "parameters": [
          {
            "name": "summary_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Summary Id"
            }
          },
          {
            "name": "topic_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Topic Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/summaries/{summary_id}/decisions": {
      "post": {
        "tags": [
          "summaries"
        ],
        "summary": "Create Decision Endpoint",
        "description": "\u8981\u7d04\u306b\u6c7a\u5b9a\u4e8b\u9805\u30921\u4ef6\u8ffd\u52a0\u3059\u308b. \u8981\u7d04\u304c\u898b\u3064\u304b\u3089\u306a\u3051\u308c\u3070 404.\n\nArgs:\n    summary_id (UUID): \u89aa\u8981\u7d04\u306e ID.\n    body (DecisionCreate): \u8ffd\u52a0\u3059\u308b\u6c7a\u5b9a\u4e8b\u9805.\n    db_session (AsyncSession): SQLAlchemy \u306e\u975e\u540c\u671f\u30bb\u30c3\u30b7\u30e7\u30f3.\n    user (User): \u73fe\u5728\u306e\u30e6\u30fc\u30b6\u30fc.\n\nReturns:\n    DecisionResponse: \u8ffd\u52a0\u3055\u308c\u305f\u6c7a\u5b9a\u4e8b\u9805 (id \u4ed8\u304d).\n\nRaises:\n    HTTPException: 404 - \u8981\u7d04\u304c\u898b\u3064\u304b\u3089\u306a\u3044\u5834\u5408.",
        "operationId": "create_decision_endpoint_api_v1_summaries__summary_id__decisions_post",
        "parameters": [
          {
            "name": "summary_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Summary Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/summaries/{summary_id}/decisions/{decision_id}": {
      "patch": {
        "tags": [
          "summaries"
        ],
        "summary": "Update Decision Endpoint",
        "description": "\u6c7a\u5b9a\u4e8b\u9805\u3092\u90e8\u5206\u66f4\u65b0\u3059\u308b. \u898b\u3064\u304b\u3089\u306a\u3051\u308c\u3070 404, \u5fc5\u9808\u3092 null \u306b\u3059\u308b\u3068 422.\n\nArgs:\n    summary_id (UUID): \u89aa\u8981\u7d04\u306e ID.\n    decision_id (int): \u6c7a\u5b9a\u4e8b\u9805\u306e ID.\n    body (DecisionEdit): \u66f4\u65b0\u5185\u5bb9.\n    db_session (AsyncSession): SQLAlchemy \u306e\u975e\u540c\u671f\u30bb\u30c3\u30b7\u30e7\u30f3.\n    user (User): \u73fe\u5728\u306e\u30e6\u30fc\u30b6\u30fc.\n\nReturns:\n    DecisionResponse: \u66f4\u65b0\u5f8c\u306e\u6c7a\u5b9a\u4e8b\u9805.\n\nRaises:\n    HTTPException: 404 - \u6c7a\u5b9a\u4e8b\u9805\u304c\u898b\u3064\u304b\u3089\u306a\u3044\u5834\u5408. 422 - \u5fc5\u9808\u30d5\u30a3\u30fc\u30eb\u30c9\u3092 null \u306b\u3057\u305f\u5834\u5408.",
        "operationId": "update_decision_endpoint_api_v1_summaries__summary_id__decisions__decision_id__patch",
        "parameters": [
          {
            "name": "summary_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Summary Id"
            }
          },
          {
            "name": "decision_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Decision Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionEdit"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "summaries"
        ],
        "summary": "Delete Decision Endpoint",
        "description": "\u6c7a\u5b9a\u4e8b\u9805\u30921\u4ef6\u524a\u9664\u3059\u308b. \u898b\u3064\u304b\u3089\u306a\u3051\u308c\u3070 404.\n\nArgs:\n    summary_id (UUID): \u89aa\u8981\u7d04\u306e ID.\n    decision_id (int): \u6c7a\u5b9a\u4e8b\u9805\u306e ID.\n    db_session (AsyncSession): SQLAlchemy \u306e\u975e\u540c\u671f\u30bb\u30c3\u30b7\u30e7\u30f3.\n    user (User): \u73fe\u5728\u306e\u30e6\u30fc\u30b6\u30fc.\n\nRaises:\n    HTTPException: 404 - \u6c7a\u5b9a\u4e8b\u9805\u304c\u898b\u3064\u304b\u3089\u306a\u3044\u5834\u5408.",
        "operationId": "delete_decision_endpoint_api_v1_summaries__summary_id__decisions__decision_id__delete",
        "parameters": [
          {
            "name": "summary_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Summary Id"
            }
          },
          {
            "name": "decision_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Decision Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/summaries/{summary_id}/action-items": {
      "post": {
        "tags": [
          "summaries"
        ],
        "summary": "Create Action Item Endpoint",
        "description": "\u8981\u7d04\u306b\u30a2\u30af\u30b7\u30e7\u30f3\u30a2\u30a4\u30c6\u30e0\u30921\u4ef6\u8ffd\u52a0\u3059\u308b. \u8981\u7d04\u304c\u898b\u3064\u304b\u3089\u306a\u3051\u308c\u3070 404.\n\nArgs:\n    summary_id (UUID): \u89aa\u8981\u7d04\u306e ID.\n    body (ActionItemCreate): \u8ffd\u52a0\u3059\u308b\u30a2\u30af\u30b7\u30e7\u30f3\u30a2\u30a4\u30c6\u30e0.\n    db_session (AsyncSession): SQLAlchemy \u306e\u975e\u540c\u671f\u30bb\u30c3\u30b7\u30e7\u30f3.\n    user (User): \u73fe\u5728\u306e\u30e6\u30fc\u30b6\u30fc.\n\nReturns:\n    ActionItemResponse: \u8ffd\u52a0\u3055\u308c\u305f\u30a2\u30af\u30b7\u30e7\u30f3\u30a2\u30a4\u30c6\u30e0 (id \u4ed8\u304d).\n\nRaises:\n    HTTPException: 404 - \u8981\u7d04\u304c\u898b\u3064\u304b\u3089\u306a\u3044\u5834\u5408.",
        "operationId": "create_action_item_endpoint_api_v1_summaries__summary_id__action_items_post",
        "parameters": [
          {
            "name": "summary_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Summary Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionItemCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionItemResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/summaries/{summary_id}/action-items/{action_item_id}": {
      "patch": {
        "tags": [
          "summaries"
        ],
        "summary": "Update Action Item Endpoint",
        "description": "\u30a2\u30af\u30b7\u30e7\u30f3\u30a2\u30a4\u30c6\u30e0\u3092\u90e8\u5206\u66f4\u65b0\u3059\u308b. \u898b\u3064\u304b\u3089\u306a\u3051\u308c\u3070 404, \u5fc5\u9808\u3092 null \u306b\u3059\u308b\u3068 422.\n\nArgs:\n    summary_id (UUID): \u89aa\u8981\u7d04\u306e ID.\n    action_item_id (int): \u30a2\u30af\u30b7\u30e7\u30f3\u30a2\u30a4\u30c6\u30e0\u306e ID.\n    body (ActionItemEdit): \u66f4\u65b0\u5185\u5bb9.\n    db_session (AsyncSession): SQLAlchemy \u306e\u975e\u540c\u671f\u30bb\u30c3\u30b7\u30e7\u30f3.\n    user (User): \u73fe\u5728\u306e\u30e6\u30fc\u30b6\u30fc.\n\nReturns:\n    ActionItemResponse: \u66f4\u65b0\u5f8c\u306e\u30a2\u30af\u30b7\u30e7\u30f3\u30a2\u30a4\u30c6\u30e0.\n\nRaises:\n    HTTPException: 404 - \u30a2\u30af\u30b7\u30e7\u30f3\u30a2\u30a4\u30c6\u30e0\u304c\u898b\u3064\u304b\u3089\u306a\u3044\u5834\u5408. 422 - \u5fc5\u9808\u30d5\u30a3\u30fc\u30eb\u30c9\u3092 null \u306b\u3057\u305f\u5834\u5408.",
        "operationId": "update_action_item_endpoint_api_v1_summaries__summary_id__action_items__action_item_id__patch",
        "parameters": [
          {
            "name": "summary_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Summary Id"
            }
          },
          {
            "name": "action_item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Action Item Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionItemEdit"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionItemResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "summaries"
        ],
        "summary": "Delete Action Item Endpoint",
        "description": "\u30a2\u30af\u30b7\u30e7\u30f3\u30a2\u30a4\u30c6\u30e0\u30921\u4ef6\u524a\u9664\u3059\u308b. \u898b\u3064\u304b\u3089\u306a\u3051\u308c\u3070 404.\n\nArgs:\n    summary_id (UUID): \u89aa\u8981\u7d04\u306e ID.\n    action_item_id (int): \u30a2\u30af\u30b7\u30e7\u30f3\u30a2\u30a4\u30c6\u30e0\u306e ID.\n    db_session (AsyncSession): SQLAlchemy \u306e\u975e\u540c\u671f\u30bb\u30c3\u30b7\u30e7\u30f3.\n    user (User): \u73fe\u5728\u306e\u30e6\u30fc\u30b6\u30fc.\n\nRaises:\n    HTTPException: 404 - \u30a2\u30af\u30b7\u30e7\u30f3\u30a2\u30a4\u30c6\u30e0\u304c\u898b\u3064\u304b\u3089\u306a\u3044\u5834\u5408.",
        "operationId": "delete_action_item_endpoint_api_v1_summaries__summary_id__action_items__action_item_id__delete",
        "parameters": [
          {
            "name": "summary_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Summary Id"
            }
          },
          {
            "name": "action_item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Action Item Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ActionItemCreate": {
        "properties": {
          "description": {
            "type": "string",
            "title": "Description",
            "description": "\u30a2\u30af\u30b7\u30e7\u30f3\u30a2\u30a4\u30c6\u30e0\u306e\u8aac\u660e"
          },
          "assignee": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assignee",
            "description": "\u30a2\u30af\u30b7\u30e7\u30f3\u30a2\u30a4\u30c6\u30e0\u306e\u62c5\u5f53\u8005"
          },
          "due_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due Date",
            "description": "\u30a2\u30af\u30b7\u30e7\u30f3\u30a2\u30a4\u30c6\u30e0\u306e\u671f\u9650"
          }
        },
        "type": "object",
        "required": [
          "description"
        ],
        "title": "ActionItemCreate",
        "description": "ActionItem\u3092\u4f5c\u6210\u3059\u308b\u305f\u3081\u306e\u30ea\u30af\u30a8\u30b9\u30c8\u30dc\u30c7\u30a3."
      },
...
      "ActionItemResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "\u30a2\u30af\u30b7\u30e7\u30f3\u30a2\u30a4\u30c6\u30e0\u306e\u4e00\u610f\u8b58\u5225\u5b50"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "description": "\u30a2\u30af\u30b7\u30e7\u30f3\u30a2\u30a4\u30c6\u30e0\u306e\u8aac\u660e"
          },
          "assignee": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assignee",
            "description": "\u30a2\u30af\u30b7\u30e7\u30f3\u30a2\u30a4\u30c6\u30e0\u306e\u62c5\u5f53\u8005"
          },
          "due_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due Date",
            "description": "\u30a2\u30af\u30b7\u30e7\u30f3\u30a2\u30a4\u30c6\u30e0\u306e\u671f\u9650"
          }
        },
        "type": "object",
        "required": [
          "id",
          "description"
        ],
        "title": "ActionItemResponse",
        "description": "\u8981\u7d04\u30ec\u30b9\u30dd\u30f3\u30b9\u5185\u306e\u30a2\u30af\u30b7\u30e7\u30f3\u30a2\u30a4\u30c6\u30e0."
      },
...
      "DecisionCreate": {
        "properties": {
          "description": {
            "type": "string",
            "title": "Description",
            "description": "\u6c7a\u5b9a\u4e8b\u9805\u306e\u8aac\u660e"
          },
          "decided_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decided By",
            "description": "\u6c7a\u5b9a\u3057\u305f\u4eba\u7269\u307e\u305f\u306f\u30b0\u30eb\u30fc\u30d7"
          }
        },
        "type": "object",
        "required": [
          "description"
        ],
        "title": "DecisionCreate",
        "description": "Decision\u3092\u4f5c\u6210\u3059\u308b\u305f\u3081\u306e\u30ea\u30af\u30a8\u30b9\u30c8\u30dc\u30c7\u30a3."
      },
      "DecisionEdit": {
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "\u6c7a\u5b9a\u4e8b\u9805\u306e\u8aac\u660e"
          },
          "decided_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decided By",
            "description": "\u6c7a\u5b9a\u3057\u305f\u4eba\u7269\u307e\u305f\u306f\u30b0\u30eb\u30fc\u30d7"
          }
        },
        "type": "object",
        "title": "DecisionEdit",
        "description": "Decision\u3092\u7de8\u96c6\u3059\u308b\u305f\u3081\u306e\u30ea\u30af\u30a8\u30b9\u30c8\u30dc\u30c7\u30a3."
      },
      "DecisionResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "\u6c7a\u5b9a\u4e8b\u9805\u306e\u4e00\u610f\u8b58\u5225\u5b50"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "description": "\u6c7a\u5b9a\u4e8b\u9805\u306e\u8aac\u660e"
          },
          "decided_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decided By",
            "description": "\u6c7a\u5b9a\u3057\u305f\u4eba\u7269\u307e\u305f\u306f\u30b0\u30eb\u30fc\u30d7"
          }
        },
        "type": "object",
        "required": [
          "id",
          "description"
        ],
        "title": "DecisionResponse",
        "description": "\u8981\u7d04\u30ec\u30b9\u30dd\u30f3\u30b9\u5185\u306e\u6c7a\u5b9a\u4e8b\u9805."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TopicCreate": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title",
            "description": "\u8b70\u984c\u306e\u30bf\u30a4\u30c8\u30eb"
          },
          "summary": {
            "type": "string",
            "title": "Summary",
            "description": "\u8b70\u984c\u306e\u8a73\u7d30\u306a\u8981\u7d04"
          }
        },
        "type": "object",
        "required": [
          "title",
          "summary"
        ],
        "title": "TopicCreate",
        "description": "Topic\u3092\u4f5c\u6210\u3059\u308b\u305f\u3081\u306e\u30ea\u30af\u30a8\u30b9\u30c8\u30dc\u30c7\u30a3."
      },
      "TopicEdit": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title",
            "description": "\u8b70\u984c\u306e\u30bf\u30a4\u30c8\u30eb"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary",
            "description": "\u8b70\u984c\u306e\u8a73\u7d30\u306a\u8981\u7d04"
          }
        },
        "type": "object",
        "title": "TopicEdit",
        "description": "Topic\u3092\u7de8\u96c6\u3059\u308b\u305f\u3081\u306e\u30ea\u30af\u30a8\u30b9\u30c8\u30dc\u30c7\u30a3."
      },
      "TopicResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "\u8b70\u984c\u306e\u4e00\u610f\u8b58\u5225\u5b50"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "description": "\u8b70\u984c\u306e\u30bf\u30a4\u30c8\u30eb"
          },
          "summary": {
            "type": "string",
            "title": "Summary",
            "description": "\u8b70\u984c\u306e\u8a73\u7d30\u306a\u8981\u7d04"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "summary"
        ],
        "title": "TopicResponse",
        "description": "\u8981\u7d04\u30ec\u30b9\u30dd\u30f3\u30b9\u5185\u306e\u8b70\u984c."
      },
...
    }
  }
}